+++ app/jobs/folio/regenerate_thumb_webp_job.rb
@@ -27,7 +27,9 @@ class Folio::RegenerateThumbWebpJob < Folio::ApplicationJob
       end
 
       if changed
-        image.update!(thumbnail_sizes:)
+        image.thumbnail_sizes = thumbnail_sizes
+        image.dont_run_after_save_jobs = true
+        image.save!(validate: false)
       end
     end
   end